cdk-java: split DestinationAcceptanceTest
